resturant/requirements.txt
asgiref==3.4.1; python_version >= "3.8"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.8"
django==4.0; python_version >= "3.8"
djangorestframework==3.13.1; python_version >= "3.6"
pytz==2021.3
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8"
